moved auth to its own routes file
