+++ client/src/pages/userform/Login.js
@@ -27,7 +27,7 @@ export default function Login({ invCode }) {
   const onSubmitLogin = async event => {
     event.preventDefault();
 
-    const response = await Client.request('/user/login', 'POST', values);
+    const response = await Client.request('/auth/login', 'POST', values);
 
     if (response.status !== 200) {
       alert('Login unsuccessful');
