+++ server/routes/index.js
@@ -3,7 +3,7 @@ const userRoutes = require('./userRoutes');
 const chatRoutes = require('./chatroomRoutes');
 const inviteRoutes = require('./inviteRoutes');
 const seedRoutes = require('./seedRoutes');
-const catchAsyncError = require('../middleware/catchAsyncError');
+const authRoutes = require('./authRoutes');
 
 const router = express.Router();
 
@@ -11,5 +11,6 @@ router.use('/user', userRoutes);
 router.use('/chat', chatRoutes);
 router.use('/invite', inviteRoutes);
 router.use('/seed', seedRoutes);
+router.use('/auth', authRoutes);
 
 module.exports = router;
